Move test mode span processor
Move the test mode span processor out of the client and with the
other span processor.
